+++ packages/cli/src/workflow-runner.ts
@@ -21,21 +21,20 @@ import { ActiveExecutions } from '@/active-executions';
 import config from '@/config';
 import { ExecutionRepository } from '@/databases/repositories/execution.repository';
 import { ExecutionNotFoundError } from '@/errors/execution-not-found-error';
+import { MaxStalledCountError } from '@/errors/max-stalled-count.error';
 import {
 	getLifecycleHooksForRegularMain,
 	getLifecycleHooksForScalingWorker,
 	getLifecycleHooksForScalingMain,
 } from '@/execution-lifecycle/execution-lifecycle-hooks';
+import { ExecutionService } from '@/executions/execution.service';
+import { CredentialsPermissionChecker } from '@/executions/pre-execution-checks';
 import { ManualExecutionService } from '@/manual-execution.service';
 import { NodeTypes } from '@/node-types';
 import type { ScalingService } from '@/scaling/scaling.service';
 import type { Job, JobData } from '@/scaling/scaling.types';
-import { PermissionChecker } from '@/user-management/permission-checker';
 import * as WorkflowExecuteAdditionalData from '@/workflow-execute-additional-data';
-import { generateFailedExecutionFromError } from '@/workflow-helpers';
 import { WorkflowStaticDataService } from '@/workflows/workflow-static-data.service';
-
-import { MaxStalledCountError } from './errors/max-stalled-count.error';
 
 @Service()
 export class WorkflowRunner {
@@ -50,9 +49,10 @@ export class WorkflowRunner {
 		private readonly executionRepository: ExecutionRepository,
 		private readonly workflowStaticDataService: WorkflowStaticDataService,
 		private readonly nodeTypes: NodeTypes,
-		private readonly permissionChecker: PermissionChecker,
+		private readonly credentialsPermissionChecker: CredentialsPermissionChecker,
 		private readonly instanceSettings: InstanceSettings,
 		private readonly manualExecutionService: ManualExecutionService,
+		private readonly executionService: ExecutionService,
 	) {}
 
 	/** The process did error */
@@ -134,10 +134,14 @@ export class WorkflowRunner {
 
 		const { id: workflowId, nodes } = data.workflowData;
 		try {
-			await this.permissionChecker.check(workflowId, nodes);
+			await this.credentialsPermissionChecker.check(workflowId, nodes);
 		} catch (error) {
 			// Create a failed execution with the data for the node, save it and abort execution
-			const runData = generateFailedExecutionFromError(data.executionMode, error, error.node);
+			const runData = this.executionService.generateFailedExecutionFromError(
+				data.executionMode,
+				error,
+				error.node,
+			);
 			const lifecycleHooks = getLifecycleHooksForRegularMain(data, executionId);
 			await lifecycleHooks.runHook('workflowExecuteBefore', [undefined, data.executionData]);
 			await lifecycleHooks.runHook('workflowExecuteAfter', [runData]);
