+++ packages/cli/src/workflows/__tests__/workflow-execution.service.test.ts
@@ -65,6 +65,7 @@ describe('WorkflowExecutionService', () => {
 		workflowRunner,
 		mock(),
 		mock(),
+		mock(),
 	);
 
 	const additionalData = mock<IWorkflowExecuteAdditionalData>({});
